oil.nvim
Tom Dixon recommended this to me. It's quickly become my favorite way of
navigating around. - is just so convenient (and it's not like i'm using
it for anything else anyways).

Adding/renaming files is a bit silly, just modify the buffer like
normal, but it'll get pretty intuitive I hope. Not like I do it often
anyways.