better configure nvim-dap keybinds
This is from playing around with the nvim-dap-ui. Holy cow is having
that UI so much nicer than before.

Logpoints are still a bit wack, because there is no wrapping in the dap
repl that it outputs to, but overall I think it's fine & good to not
have to re-launch Electron just for logging changes.